fix tests
Or at least, of what passes as some ...
