--- test/suite.rb
@@ -1,1 +1,0 @@
-require './test/test_bundles'
